+++ code/latte/integration/PolyRep.cpp
@@ -36,7 +36,7 @@ void _parseMonomials(_MonomialConsumer<RationalNTL>* consumer,
 	int *exponents = new int[varCount];
 	RationalNTL coefficient;
 
-	for (int i = 1; i < line.length() - 1; i++) //ignore outermost square brackets
+	for (size_t i = 1; i < line.length() - 1; i++) //ignore outermost square brackets
 	{
 		if (line[i] == '[')
 		{
@@ -170,7 +170,7 @@ void _parseLinForms(_FormSumConsumer<RationalNTL>* consumer, const string& line)
 	int degree;
 	RationalNTL coefficient;
 
-	for (int i = 1; i < line.length() - 1; i++) //ignore outermost square brackets
+	for (size_t i = 1; i < line.length() - 1; i++) //ignore outermost square brackets
 	{
 		if (line[i] == '[')
 		{
