+++ code/latte/valuation/valuation.cpp
@@ -1252,7 +1252,7 @@ void Valuation::ValuationContainer::add(const ValuationData & d)
 void Valuation::ValuationContainer::printResults(ostream & out) const
 {
 	out << "\n";
-	for (int i = 0; i < answers.size(); ++i)
+	for (size_t i = 0; i < answers.size(); ++i)
 	{
 		if (answers[i].valuationType == PolytopeValuation::volumeCone)
 			out << "Volume (using the cone decomposition method)" << endl;
